--- Demo/Assets/GoogleVR/Prefabs/Utilities/GvrFPSCanvas.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6dfff844c95f342d1a8a9ab165303662
-timeCreated: 1471563509
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
