--- Demo/Assets/GoogleVR/Scripts/Utilities/GvrFPS.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 644076fb2b4854a72b6de9578e358b8e
-timeCreated: 1471565344
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
