+++ Demo/Assets/HUD stuff.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 6d6413db92f36db4593cb61994f1ba35
+folderAsset: yes
+timeCreated: 1493309364
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
