+++ Demo/Assets/Scripts/vrHUD.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 091668259d53a3b49a78b312654f454f
+timeCreated: 1493311258
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
