+++ Demo/Library/LastSceneManagerSetup.txt
@@ -1,4 +1,4 @@
 sceneSetups:
-- path: Assets/Scenes/TableScene.unity
+- path: Assets/Scenes/WorldScene.unity
   isLoaded: 1
   isActive: 1
