--- Demo/Library/ScriptAssemblies/BuiltinAssemblies.stamp
@@ -1,2 +1,0 @@
-0000.58dcb854.0000
-0000.58dcb86a.0000
